Laying foundations for my homepage
Created the header and navigation place holder with the gallery and the mini description box paired with the "Learn More" button and image of a solar house to the right of it.

As of right now, the website has no true functionality. Just coloured boxes saying where is where and giving an idea of what is going be their. I did have to learn house to use 'Flex' on CSS as I struggled with strecthing the 'Learn More' button the the bottom of the 'descTxtBox' divider.

I used many percentages here for keeping in mind for resizing. This is not the final design of the website. All of this is just prototyping.