FINALLY, I some how managed to fix this bug
Had to do some research on how 'position' works but now my website should work for extreme compression and work for the very common 1920x1080 montors up to around 2000 pixels in width. I cant get this website to work for more than that because my laptop cant actually produce that for some reason but I am proud of this resizing ability.